Revert "Better preload drawable logging"
Bug: 63136392
Bug: 150395371
Test: make

This reverts commit cc7258f38013a9f6326c7d28b7ca11f4c1bba929.

Bitmap moving into a new mainline module in S. These module-boundary-
crossing hidden APIs need to be either made official or removed.

We already got the information we need from this logging, so remove it
and the hidden APIs. If/when we need to investigate this in the future,
we should be able to recreate the logging without crossing the boundary.

Change-Id: I5d094a0e427d225719614745867d5b00888d8fec