RebootEscrowManager: always report success metric
In the process of OTAing from a build that does not have the armed
status marker, all metrics are going to be lost. Since a success is
a sure indication that reboot escrow was prepared before reboot, then
we should mark it as a success always.

On the other side, a failure could be a manual reboot if we did not
have the armed marker.

(cherry picked from commit bdb3cdc49741b0eff3ebc138f1771baae17638c5)

Bug: 149833390
Test: atest RebootEscrowManagerTests
Merged-In: Iaa179039eb7f587a6596a70176d42761e09b0fac
Change-Id: Iaa179039eb7f587a6596a70176d42761e09b0fac