Reset controls when playback state is NONE
If a notification is cleared it won't always send an update first and
call onSessionDestroyed. NotificationMediaManager already checks if the
current media notification is being removed and sends an event, so we
can listen for that.

This fixes the bug with cast notifications. It can also be tested by
playing media, pausing, and then clearing all notifications - the
controls will now switch to the resumption state.

Also removed the long-press menu in QSMediaPlayer (which was just for testing)
since I was refactoring that block anyways.

Fixes: 150437753
Fixes: 150742919
Test: manual
Change-Id: I0ecada16967a6d21b82dc86d05544cf78027bcd9
(cherry picked from commit 3316d251d322f80a8e24f0b1b2d0747f928ae6de)