Inline WebViewUpdater into WebViewUpdateServiceImpl.
Since the removal of the fallback logic and other complexity required by
Monochrome, WebViewUpdateServiceImpl did not have much logic left in it
other than forwarding most method calls to WebViewUpdater. The division
of logic between these two classes was thus mostly just making it harder
to understand.

Therefore, we inline the entirety of WebViewUpdater into
WebViewUpdateServiceImpl, combining them back into a single class as
they used to be when they were simpler. A number of methods are made
private as they don't need to be accessed from outside the class any
more.

Bug: 129470358
Test: atest WebViewUpdateServiceTest
Change-Id: I1cef1c89f47a3ddd980d0f7d27a09e9a1e1b913b