Adjusting focus starting from sibling rootable app tasks
The next rootable app task was not moved to front while
another rootable app task which did not have running activities
was on top. The task focus was not adjusted properly
because the focus was adjusted between root tasks only.

In that case, the next task info was delayed to dispatch
to organizer until the finishing activity on top was destroyed.

Bug: 152408408
Test: atest ActivityRecordTests
Change-Id: I4ce5bf4ee3a48155654cc300eb51da5e2262962e