Special-case backup/restore of replaced settings
Bug: 153940088
Test: atest SettingsProviderTest:SettingsHelperTest

Values for some settings might be changed temporarily. If a backup happens at that moment, we want to backup the real values instead of the temporary ones. If a restore happens at that moment, we don't want to restore values for modified settings. See https://b.corp.google.com/issues/153940088#comment2 for context.

Change-Id: I4866f56376ffa393220bbef828a4b876d586146b