Clear fixed rotation if display config does not change
There is a timing that when remote rotation is completed, the
orientation from sensor has been updated again. If the transition
of rotated app hasn't complete, the display is still in previous
orientation, then if it is the same as the last value from sensor,
the display configuration does not change. So in this case the
rotated app needs to be restored to avoid showing different
orientation than the display.

Fixes: 153327533
Test: ActivityRecordTests#testActivityOnCancelFixedRotationTransform
Test: Add delay to execute DisplayRotation#continueRotation.
      Put device in landscape with a portrait home on top, launch an
      activity without fixed orientation, and then rotate the device
      to portrait immediately. The activity should launch in landscape
      and rotate to portrait with animation.

Change-Id: Ic966923a751d94cc6ea86a83c2f600424999799c