Create a channel using which PMS can return result of verification
In the existing flow, once session is verified, it gets installed
immediately. In future, once a session is verified, we want to perform
some actions before we move onto installation phase.

Note that we only return the verification result, but do not stop the
installation or process the result in any way yet.

We still don't have any easy way to resume installation from where
verification ends. This will be created soon.

Note that, in order to create this channel, I had to temporarily add two
observers to InstallParams. This is because it now needs to return
verification complete notification AND result of installation. In
future, we will create a duplicate of InstallParams called
VerificationParams to handle verification exclusively and this temorary
observer will be deleted.

Bug: 159331446
Test: atest StagedInstallTest
Test: atest GtsSecurityHostTestCases
Change-Id: I8aa4a72ebbf4fa1616d0be70343f9c4f9cb98f87