Disable shared library support for framework modules
Defaults shared library to false as the majority of framework modules
should not be usable as a shared library. The exception is ike but it
does not currently use the defaults and will override this setting
when it does.

(cherry picked from 15ad568af0dc6d8290955fb44912de8270dc912d)

Bug: 156723295
Bug: 155164730
Test: m droid and flash
Change-Id: I3e78994a9855110b02c3d71944ff38b986208e9a
Merged-In: I3e78994a9855110b02c3d71944ff38b986208e9a