Updates the icon of magnification button by magnification mode
When the current magnification mode is fullscreen mode,
magnification button will use "open in new window magnification" icon.
It means that the current magnifion mode will be changed to window mode
when clicking the button in full screen mode.
And vice versa for the case that the current mode is window mode.

Bug: 145780606
Test: none
Change-Id: I4a4b082465546135afa36f1751bcf4c85e78b1c4