Plumb injector through all location code
Adds plumbing to utilize the new injector through all location code,
adds new fakes to be used for testing with the injector, and adds a
new utility helper to be utilized in future CLs.

For each *Helper.java, the abstract interface has been extracted out
to *Helper.java, the normal implementation to System*Helper.java, and
the test implementation to Fake*Helper.java.

Test: manual + presubmit
Change-Id: I55172f75d21ee55e6b58442727ddcbc1213395d3