Do not animate NSSL when in battery saver
Until now, we were only checking if the screen shold turn off on the
DozeMachine. This works but doesn't propagate the always-on state
properly to other parts of SysUI.

It's a better idea to hide the implementation detail under
DozeParameters, which is a singleton that's shared with other parts
of the system.

Test: manual
Test: atest DozeParametersTest
Fixes: 158771494
Change-Id: Ifd43b9b1ea247f4e869505f02d6716400de35091