Fix a provider bug and remove redundant GNSS logic
This fixes a bug where screen on/off state changing was not refreshing
the active state of registrations. In addition remove code from
GnssLocationProvider that is now redundant (all the throttling logic is
handled in the manager now).

Finally this removes a broken test from the fused provider tests. This
was broken by the last change because FusedLocationServiceTest is not
allowlisted for being able to use bypass requests (to be honest, I'm not
sure why this test was passing previously...). FusedLocationServiceTest
is not a great test, so I am removing it to fix tests for now while I
consider a better approach.

Bug: 162949881
Bug: 162950120
Test: manual + presubmit (atest BatterySaverLocationTest)
Change-Id: Ide762b352014da4aa42959d479f2e45875ec19af