Revert the fullscreen flag check removal for layout with cutout
This is a quick partial revert of the patch 353f4e2 to fix the
inconsistency after the clean-up work.

The camera team reported that with the clean-up, their in-app button
layout get messed up. Nothing should be affected by the clean-up work.
Thus, this CL is to revert the FLAG_FULLSCREEN check removal in
layoutWindowLw to keep the behavior the same.

The flag will be removed after we figure out why it is not compatible
with the app.

Bug: 167425558
Test: See reproduce steps in the bug.
Test: atest DisplayPolicyLayoutTests
Test: atest DisplayPolicyInsetsTests
Change-Id: I150d82dc5a233e7e7f458b83a379b1fa1770f0ea