HexDumpTest: Fix assertion.
Commit 1e165f1d440e85d84c6db81fe46ca02c206f611e added test coverage,
including a helper function assertThrows(). The implementation of that
method was buggy because it only failed the test if it was throwing the
wrong kind of RuntimeException, instead of also when the code under test
wasn't throwing an exception at all.

This CL fixes the helper function. The test still passes (none of the
code under test was relying on the incorrect assertion).

Test: atest FrameworksCoreTests:com.android.internal.util.HexDumpTest

Change-Id: I8fe3e70ada8b880a03e124cb48e204e7bb880e57