Add AuthBiometricUdfpsView for portrait
Adds AuthDialog.LayoutParams which should be used to hold any
layout or measurement related information.

Adds onMeasureInternal which returns AuthDialog.LayoutParams
so that subclasses can return custom layout parameters while
keeping common things like setMeasuredDimension in the
base class.

The AuthBiometricUdfpsView in portrait mode includes a
spacer whose height is calculated based on the sizes of
everything that's below the physical sensor, including
1) Text indicator area (for error messages)
2) Footer button bar
3) BiometricPrompt distance to navigation bar
4) Height of navigation bar

The indicator text is the only non-static element below the
icon. Changes it to marquee on a single line for now.

AuthBiometricView animates the icon's holder instead of the
icon itself. Otherwise, the holder would need to allow drawing
outside of its bounds.

Test: No effect on existing devices
Test: atest com.android.systemui.biometrics

Fixes: 171071087

Change-Id: I3b62f02a44f4d3cf6925d7782451596fbf8db0e1