Fix the focus not move to the place where the finger clicks
User does the triple tap operation, but the gesture detection
becomes to doubleTapAndHold because the timing of the motion
event. Then the long pressing data is created but not be cleared
when the state returns to STATE_CLEAR.
Adding the long pressing data reset to fix this issue.

Bug: 162521649 and 163298552
Test: a11y CTS & unit tests
Change-Id: I40d89ea958bee619c64799de591be96bf92efbfc
Merged-In: I5fc23a14b0a4ae3c7c012cb4ed8dc924c00e7fe9