Add new carrier config option for enabling SIP PUBLISH of RCS caps
Adds a new carrier config option that will bring up/down the service
connection to the vendor IMS stack. When down, we should not
see the vendor IMS stack generate SIP PUBLISH requests to the network.

This is so that we do not generate SIP PUBLISH requests on networks
that do not support this functionality.

Bug: 173115560
Test: atest CarrierConfigTests
Merged-In: If9ab9836261afc2d6cebb959690d8b44b06a3dfe
Change-Id: If9ab9836261afc2d6cebb959690d8b44b06a3dfe