Add notion of Quirks the Device Display Config files.
On some devices, the system may not operate as expected and
specialized code may be necessary to work around these limitations
or "quirks".  Add the notion of quirks to the config files so that
devices can be easily configured to operate around their
unexpected behavior.

As a first "quirk" to add, we specify if a display supports setting
brightness via the SourceControl.setBrighntess API.

Bug: 170954431
Test: manually check that brightness works
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/display
Change-Id: Ia3cedf9ec7a2457e27763e0c58020c1edfbae408