Correctly update expedited job connectivity constraint.
Expedited jobs punch through firewalls due to their elevated procstate.
However, the proc state isn't elevated until the job starts, so we can't
wait for a callback from the networking stack to mark the connectivity
constraint as satisfied (as long as the network exists). This
firewall-hole-punch-after-elevated-procstate is the same behavior that
foreground services see, so expedited jobs will get the same treatment.

Bug: 176922022
Bug: 171305774
Test: atest CtsJobSchedulerTestCases
Test: atest FrameworksMockingServicesTests:ConnectivityControllerTest
Change-Id: I844a5ec2157f1ec3cee1574584eea2960e1a3b62