Migrate legacy intent filter verification settings
These legacy settings need to be migrated once to the new domain
verification API, so this pulls the parsing out of the legacy classes
into a DomainVerificationLegacySettings.

This does NOT handle serializing the legacy settings. It's assumed
that migration is a one-time best effort attempt, and that dropping
the legacy settings is an acceptable fallback.

The old setting was a result of an unexplained and unintuitive user
path, set through ResolverActivity, and so it's not clear whether we
should migrate at all.

Worst case, the user can always re-do their preferences.

NOTE: This change should only be merged if merged together with the
change that removes the legacy code. It's invalid to run these
separately, as this change breaks the legacy manager class. These are
only separate to make review easier, to separate their concerns.

Exempt-From-Owner-Approval: Already approved by owners on main branch

Bug: 171251883

Test: TODO

Change-Id: Iefde9f2cd8ab73cb5f592abc7d2163dd7d244789