New Bubbles animations!
Turns out we can animate SurfaceView alpha if a) it's Z ordered to the top, and b) we use a ValueAnimator to call .setAlpha() specifically - using .animate().alpha() sort of works but results in weird flickering. Go figure!

This CL uses that newfound ability to add:
- Fade/scale down the expanded view when you drag out the expanded bubble
- Softer expand/collapse animations
- Softer transition between bubbles
- Fixes for flashes of old bubbles while transitioning between bubbles/overflow

Will follow up with polish to exactly match the mocks, but this gets the big underlying changes in!

Test: atest SystemUITests
Test: expand and collapse bubbles
Test: drag out bubbles/dismiss the currently expanded bubble
Test: switch between bubbles of different heights
Test: all of the above one after another
Fixes: 170268335
Fixes: 177930422
Fixes: 169595742
Change-Id: I5913c05392050a4ab736d34c5cea935bf61d3db8