Use different prompt messages for biometric|credential
Because the subtitle/description string is shared between the biometric
and credential versions of BiometricPrompt, the corresponding prompt
message string should indicate that either auth type may be used.
Otherwise, the string "Use your fingerprint to continue" may be shown on
a screen prompting the user for their PIN, or vice versa.

This commit addresses the issue by having the getPromptMessage(int)
method of BiometricManager return a different string when both biometric
and credential authentication are allowed and available, as opposed to
when only biometric authentication is allowed.

Test: atest BiometricManagerTest
Test: Manually tested strings using biometric integration test app

Bug: 180732913
Change-Id: I39d22c756b71de533f228b6a765ef81b82c2f4da