Remove configuration of com.android.geotz module
This commit removes system server configuration associated with the
"offline" location time zone provider from the nascent
com.android.geotz module, which has been configured as the "secondary"
provider. This module will not to be updatable in S so needs to be
removed before release.

This is effectively a revert of commit
475e9ba0039cbc853ffb81b21376b7830a721f26 with adjustments for changes /
extra references that have occurred since.

Bug: 178169250
Bug: 172546738
Test: treehugger
Change-Id: Ieeeb2891c6a17cc06bed48e6849a75ebdd29e28f