Log ThemeOverlayController debug logs
- We had some reports of theme colors changing unexpectedly.
- We believe this is likely addressed by avoiding creating different
color schemes based on if the user is in dark mode and light mode at the
time the scheme is generated. This doesn't work well because the user
will eventually exit the current light/dark mode and be stuck with the
the 'wrong' colors
- Enabling debug logs will let us have more data
if the bug reproduces on recent builds, as it will show what triggered
generating new colors.
- We expect low additional traffic from these logs, as they only occur
when new colors are generated, and there are very few events that
trigger generating new colors, and those events are relatively rare
(ex. on boot, assigning a new wallpaper, after completing work profile
setup)

Test: none
Bug: 183392558
Change-Id: I503d9ec88fdc5f84418197bbde6cead59454e40d