Fix intermittent async noteOps
This removes the 'sThreadsListeningForOpNotedInBinderTransaction'
variable and its usages.

I think the intention of this variable was to act as a bloom filter:
i.e., "threadId is possibly in set or definitely not in set", but...
it's actually only giving us "threadId is possibly in set or possibly
not in set". This leads to false negatives, which leads to incorrect
behavior.

In this case, the incorrect behavior was that noteOps would sporatically
turn into async noteOps.  This is currently causing a few tests to be
flaky: when setting a setOnOpNotedCallback listener, sometimes noteOps
code come back sync, sometimes async, and sometimes they get lost in transit.

Bug: 187722787
Test: atest --iterations=100 AppOpsLoggingTest#getBTScanResults
Change-Id: I26c7b8029ea140d850aba7ec77c691ed9fea76d9