Break VolumeDialogControllerImpl' dependency on StatusBar
VolumeDialogControllerImpl had a reference to a lazy StatusBar.
Because of how it was used, this could result in trying to stand up
StatusBar (and all of its dependencies) on a non-main thread, which
they did not appreciate.

This removes the dependency on StatusBar, and throws in a healthy
dash of dependency injection while we're at it.

Fixes: 189203821
Test: atest SystemUITests && manual
Change-Id: I41e00438b30928815e9355162af05f926f24dfce