Add base Agent code.
The agent handles the nitty gritty details of TARE, interacting directly
with ledgers, carrying out specific events such as tax collection and
granting initial balances or replenishing balances, and tracking
ongoing events.

The base code just handles granting ARCs at setup, app install, and
charging, and reclaiming ARCs on app uninstall. Transactions for
instantaenous and ongoing events will be added in a followup as there
are additional dependencies that need to be checked in.

Bug: 158300259
Test: Android builds
Change-Id: I639c2b74abb32fc7867a2b0e8d2bc8e99a7c41f4