Skip close activity transition since no closing app can be visible
The layer of transition animation is usually shift to animation layer,
so when transition controller going to apply a activity close
transition to a bottom activity, the closing animation surface will be
on top of the task, then suddenly occludes the top activity.
An easy fix is to skip the close activity transtion if there is no
closing activity can be visible.

Bug: 186860966
Test: atest AppTransitionTests AppTransitionControllerTest
Change-Id: I3869530fc1eaa87ed1403fed86ff68a1e4de5b23