Update wallpaper offset when changing target from null
updateWallpaperOffsetLocked copies the IPC received values to the state
variables that are actually set on the wallpaper window. It skips
this job when there is no wallpaper target. In a situation where we are
going from no wallpaper target to wallpaper target (e.g. lock screen
wallpaper) there may be nothing to trigger this method to be called
again when we finally receive a wallpaper target, and we then have
to wait for another IPC. Since there were some values ignored, it
can lead to some jank/misalignment.

Bug: 195596603
Test: Repro from bug.
Change-Id: I3ef19b77bc33546a3e80bca75532d017b4712054