Return the correct root task when start activity from recents
When starting activity from recents, the system will find the
suitable task and move it to the front.  With corresponding task
display area, TDA#getOrCreateRootTask returns the candidate task
directly instead of the existing launch root task.
In RootWindowContainer#anyTaskForId, the candidate task is reparent
to the root task(same as itself) which resolve from the
TDA#getOrCreateRootTask and lead to the watchdog timeout. This CL
updates the logic to return the correct root task.

Bug: 197293995
Test: atest MultiWindowTests,WMShellUnitTests
Change-Id: Ib0b24363dc35d624c358080998c25874d9e7094c