Fixing tiles losing subtext in split shade
The issue was caused by rotation from shade with fewer number of visible QQS tiles (portrait mode) to more visible QQS tiles (split shade). For tiles that were not visible in QQS, QSAnimator was setting their respective QS tiles alpha to 0. This value was not updated when they become visible in QQS split shade, so this change is making sure that all QS tiles that have corresponding visible QQS tile will have alpha set to 1 at the start of the animation.

Bug: 196831963
Test: open portrait shade -> rotate to split shade -> expand QS - all tiles should have subtitles
Change-Id: I44b2a532232748ece926d6ac7819ec0b8a809a51