Add required flag to registerReceiver calls for platform services
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
receiver should be exported for apps targeting T+ that are registering
for unprotected broadcasts. This commit updates all of the calls under
services/ that are registering receivers for unprotected broadcasts.

Bug: 161145287
Test: Manually verified code did not trigger logcat message for not
  specifying the necessary flag when calling #registerReceiver
Change-Id: Icaef9d9cf9ed888d124f33c40dbeb1b15df87938