Letterbox: Add rounded corners above a taskbar.
Previously rounded corners were applied to the whole app window regardless of the taskbar insets which resulted in bottom rounded corners being hidden behind the taskbar. This change fixes that by cropping activity surface at the taskbar top.

Also, starts to determine corner radius automatically based on the device corners radii but leaves an option to override it with a config overlay for flexibility.

Fix: 197549955
Test: atest WmTests:SizeCompatTests and manual
Change-Id: Id687858bca3277f9f680f78d19c4fbc86d192b41