Don't show nested badge for work profile direct share targets
Currently, we add a work badge (briefcase icon) to app icons in
the sharesheet. If the target is an app, this results in the
regular app icon with a work badge. However, for direct share
targets we show the person, with the app icon as a badge. This
means that for direct share targets, we show an app icon badge
that itself has a (very tiny) work profile badge.

This change removes the "nested" work profile badge for direct
share targets.

Bug: 179783660
Fix: 179783660
Test: manual (shared something and verified that no direct share
targets have a work profile badge, but work profile apps are still
badged)

Change-Id: I5196184a6f1fb442eee905240656faa6c9fc4cfb