Fixed data icon disappeared issue
When carriers have more than one APN that supports
internet APN type, the data icon might sometimes
disappeared, for example, after sending an MMS.
This is caused by the PreciseDataConnectionState
refactoring in Android 12. Added the workaround in
TelephonyRegistry to aggregate the data connection state for
internet, since it would be very difficult to make this
right in DataConnection class.

In Android 13, a more robust and comprehensive logic will
be added in the new data network controller to fix this.

Fix: 204328662
Fix: 187967526
Test: Manual test & TelephonyRegistryTest
Change-Id: Icba446e879e4d892829b41fdea50ebc1a10bf832