Assign layers after adding a window
When adding a new window we create a WindowToken
that initially created without a surface.
When this window is attached to parent
we call onParentChanged that tries to assign the layers
for correct z-ordering of the surfaces
but this call is ignored for the newly created
WindowToken as it doesn't have the surface yet.
The surface for the WindowToken is created only when
the 'window' child is added but currently we do not
update z-order for it (it may happen only on the next
relayout).

Added logic that invokes assigning of layers again when
a window is added and when surface is ready.

Bug: 200103245
Test: add rotation suggestion button window, check that it is displayed above taskbar
Test: atest WmTests:WindowTokenTests
Change-Id: I38c2c4416edadd346c80d871325aa029b2c1445f