Don't show 'try fingerprint' msg on face auth help
Sometimes multiple biometric-help suggestions for face auth
can be sent right before a face auth sucesss. If udfps
is suppressing the face auth help message, don't show any messages
(previouosly we'd show "try fingerprint" which was unhelpful
and sometimes was shown on screen as a successful face auth
was occurring which was confusing).

Note: we will continue to announce the help
messages for a11y purposes.

Test: atest KeyguardIndicationControllerTest
Test: manually auth with face, see keyguard message
Fixes: 203897784
Change-Id: I75b9e19b812505ff2dafa960600da67135448042