Do not allow null signal thresholds in SignalStrengthUpdateRequest
SignalStrengthUpdateRequest is designed to not allow null nor empty
signal thresholds for public apps.

To support system privileged app (Bluetooth) which is only interested
in requiring to report system thresholds while idle. An exemption was
made (aosp/1569861) which was to allow empty thresholds but the null
thresholds was also allowed by mistake. This will cause Bluetooth
crash if no signal thresholds are set.

To fix the issue, we only allow Bluetooth to set EMPTY thresholds
but not null one.

Bug: 205585585
Test: atest SignalStrengthUpdateRequestTest
Change-Id: I245ecb87ab807a6005b32fa320c8820a2c0b1a9e