Game Loading Time improvement by boosting CPU via ADPF
Adds the concept of Game State (supplied by the application) to the
GameManager API. This allows the setting of various modes to describe
the current high-level runtime state of the game, as well as the loading
state. In Performance Mode, during loading a hint is sent to the device
PowerHAL via the PowerManager.

Bug: 201769701
Test: atest android.gamemanager.cts.GameManagerTest
Ignore-AOSP-First: Mentions a confidential API
Change-Id: I695e45b12f9a30c1824df210bfefbd5a9b4bd935