Make screenshot sound respect system mute
In some regions (e.g. Japan) we force the camera to make a sound
regardless of the user settings. The screenshot used the same
sound so inherited this behavior. This change forks the sound used
for screenshots, so that it respects the system volume instead of
being forced audible.

Bug: 200767953
Fix: 200767953
Test: manual (made AudioService.readCameraSoundForced return true,
then test before and after change to verify behavior fixed)

Change-Id: I3c9cf2826b5c3f86e47b570f75189405845a7186