Avoid removing USB state changes during CONFIGURED uevent
In order to avoid missing the CONNECTED uevent, skip removing previous
events if moving into the CONFIGURED state.  The Tethering class
relies on the CONNECTED state change to tear down the USB RNDIS
adapter.  If this is skipped, the Tethering class will see no change
to the USB state, and avoid restarting of the dependent tethering
services.

Bug: http://b/207317644
Test: USB RNDIS IP assigned to PC
Change-Id: Iab800ee6b995f1ed8bb1f98934245dd010f4e560