Clean up no-op methods of display configuration
The parameter newConfiguration of
RWC#setDisplayOverrideConfigurationIfNeeded is always the reference
of displayContent.getRequestedOverrideConfiguration() which is from
DC#onRequestedOverrideConfigurationChanged (since commit 09e1af1).
It means that the method always diff the same config instance, which
is no-op. The concept of "Override configuration of the default display
duplicates global config." is still done by DisplayContent#
updateDisplayOverrideConfigurationLocked. So the methods on the path
of WMS#setNewDisplayOverrideConfiguration can be removed.

Bug: 163976519
Test: atest DisplayContentTests#testDefaultDisplayOverrideConfigUpdate
Change-Id: I6d6f44d1a5ef0c8180b1dbbc4755b0299d65dd4d