Send Task Fragment Info Update When Task Invisible...
...only when the Task has no more running activities.

Send a TaskFragment info changed callback if the
callback is for the last activities to finish in a
Task so that the TaskFragmentOrganizer can delete
this TaskFragment. Otherwise, the Task may be
removed before it becomes visible again to send this
callback because it no longer has activities. As a
result, the organizer will never get this info changed
event and will not delete the TaskFragment because the
organizer thinks the TaskFragment still has running
activities.

Bug: b/214090984
Test: atest TaskFragmentOrganizerControllerTest
Test: atest ActivityEmbeddingLaunchTests
Change-Id: Idb58975f4bfb6c67ed6cda3dc6b7c33abd960bbd