Check current user when the operations runs for fingerprint hidl.
The current and target user are both provided to the switch
client constructor which is brittle unless operations are scheduled and
run synchronously. Commit 44325741562dd7bc2c5b11ec7075a585eee7e684
changed the handlers and somehow caused a bug where the cached current
user was out of sync. The switch client tries to optimize for this and
can skip switching when that occurs.

This also includes two additonal changes 1) a few updated logs from
the original change and 2) restores each scheduling having its own
handler for clarity.

Bug: 213962104
Test: atest UserAwareBiometricSchedulerTest BiometricSchedulerTest BiometricSchedulerOperationTest  SensorTest Face10Test
Test: manual (flash, enroll fingerprint, add work profile account, reboot, verify fingerprint still works)
Change-Id: Ifb73b0145aeb8afb62d1f55d2f881347b0d2ef8a