Handle channel changes for bubbles
Bubbles in the overflow might not have an active notification associated
with them, this means if the user disables bubbles for that conversation
through settings we wouldn't be able to know and remove the bubble from
the overflow.

This change pipes the onNotificationChannelModified signal from
NotificationListenerService through sysui & to bubbles for both the
pipeline & NEM version & uses this signal to remove matching bubbles
from the overflow if appropriate.

Test: BubblesTest NewNotifPipelineBubblesTest NotificationEntryManagerTest NotifCollectionTest
Bug: 171235607
Change-Id: I3d4889ece13d550e0cceb20036718faee4887489