Inline InputMethodMenuController.HardKeyboardListener
This is a follow up CL to our previous CL [1], which introduced

   InputMethodManagerService.HardKeyboardListener,

which was later moved to

   InputMethodMenuController.HardKeyboardListener

as part of refactoring [2].

This CL mechanically inlines it to InputMethodManagerService as it
allows us to achieve the same behavior with fewer lines of code.

This is mechanical refactoring. There should be no observable behavior
change.

 [1]: Ica768083f95c33dc1e494a28ba7d8b6eb989b0ef
      01038492ff0317f0d3cff54d8a7ee36bb31ff175
 [2]: Ief3c691029b14d848f6ad4a5ec8a8743760c7cda
      adc5ef3d03b2492d728976cc43ae2b1199d3e459

Bug: 192412909
Test: Manually verified by attaching/detaching a hardware keyboard.
Change-Id: I9cce75011a8564c9da8307c5555d87bcba4db2cb