Remove HardwareAuthToken parameter from unlockUserKey
Due to the migration to synthetic passwords, the 'token' parameter to
unlockUserKey() is no longer needed.  Remove it.

Note: I didn't change unlockUser() in IActivityManager because it is
marked with UnsupportedAppUsage, so it might not be safe to change the
method signature.  It now just ignores the 'token' parameter rather than
passing it down the stack.

Test: atest com.android.server.locksettings
Bug: 184723544
Change-Id: I35ce09412f47f2f2a17a371d518a0a518b70bfb6