Also require IME visibility in isImeRenderingNavButtons
This is a follow up CL to my previous CL [1], which attempted to hide
the back button when

  InputMethodService#canImeRenderGesturalNavButtons()

returns true and the gestural navigation is enabled.

As shown in Bug 217668258 it turns out that SysUI still needs to show
the back button at least on the setup wizard.  This CL aims to avoid
such unexpected breakages by also requiring the IME visibility when
determining isImeRenderingNavButtons.

 [1]: I7a2e9f08072a97d61b78dddb144282a5728d1407
      8e005126f5aef99ba77c62e1ebd88791cf82469e

Fix: 217755059
Test: Manually verified as follows.
  1. Build coral-userdebug and flash it.
  2. adb root
  3. adb setprop  \
      persist.sys.ime.can_render_gestural_nav_buttons true
  4. adb reboot
  5. Make sure that the back button is shown on the setup wizard.
  6. Complete the setup wizard.
  7. Tap any text input field.
  8. Make sure that Gboard is shown up.
  9. Make sure that the back button is shown.
 10. adb shell dumpsys input_method | grep mNavigationBarController
     -> Make sure it's not null.
Change-Id: Ib48e1e3c9033408e5eb26f0aaf957b5c62308053