Spatializer: more SysApi for availability, head tracker
  Add @SystemApi to check if the Spatializer will be available for
a given device, if it has a head tracker, if enabled, and
change its tracker enabled state.
  Implement checking device compatibility with transaural and
binaural modes. Create list of compatible devices, list of
devices that have been enabled, and list of disabled. This is
to support the V3 UI where add/removeCompatibleDevice is used
to control Spatial Audio state for a subset of output devices.

Bug: 210803914
Test: atest SpatializerTest
Change-Id: I0acf0c0310dbd503de2ea87328fca7e5597ae32d