ColorUtils: Fix out-of-gamut Oklab blending
It's normal for some colors to go out-of-gamut when blending in Oklab,
and slight hue/lightness shifting is fine for our purposes, but failing
to mask 8-bit R/G/B channels results in broken colors when a channel is
negative or >255. For example, this can cause the lock screen <-> AOD
clock transition to flash purple when the wallpaper is saturated green.

Clamp each color component to 0-255 to fix the issue.

Change-Id: I4d6420ee5be365d7cf89d079b1b4ddf0a7102641
Signed-off-by: romiyusnandar <yusromi04@gmail.com>