[Shell-transition]: More checking animating state methods
As IME snapshot will be attached during the IME target window
in app transition or quick-switch, it still requires in WM core
since IME visiblity/insets state is controlled by WM.

As the reason, with migrating to Shell-transition, WM needs an
alternitive way to check if the IME target is transitioning by
app-transition or recents animation.

This CL introduces TransitionController#inRecentsTransition and
WindowContainer#inAppOrRecentsTransition utility method as for
IME snapshot mechanism preparation.

Bug: 201139553
Bug: 212570341
Test: build
Change-Id: Icd702ff47859ca2f0982b877177098b304ff3b8c