Keyguard was hiding occluding activities
Most reproducible while the device is asleep on AoD for a period of
time. The system detects the double tap power gesture, and launches
the camera activity immediately. However, in the failing case,
DreamManagerService only stops the dream after the activity has been
launched. This setOccluded path has already happened when the call to
wakeup comes in and decides that the StatusBarState should equal
KEYGUARD, which is incorrect. This blocks the activity which is
running underneath. The correct state is SHADE. Add a check for
occlusion before deciding to show the keyguard

Fixes: 199349731
Test: launch camera over lockscreen
Change-Id: I817fcdc80db79ad95367dc7a49eadaa5e366c3ef