Update IME SECURE flag if Window updates SECURE flag
If the IME target is set before the IME target updates its SECURE flag,
the IME  may not get the SECURE flag. This is because we only update the
IME flag when the new target is being set.

This change ensures we also update the IME secure flag if needed when a
Window updates its SECURE flag since it could already be the target.

Fixes: 215005011
Test: testImeSecureFlagGetUpdatedAfterImeInputTarget
Test: Add SECURE flag after IME target is set.
Change-Id: Ie59042c8efa2527208e4020107b7aa15c885755f