Remove framework-connectivity-proto from platform
In the generated javastream and cppsteam incident proto file,
it defines the start position for netstats at 3001 and it doesn't
actually use the NetworkStatsServiceDumpProto which is deinfed in
netstats.proto.
Thus, the framework-connectivity-proto can be removed from
framework-javastream-protos and framework-cppstream-protos.

The generated incident proto file defines netstats as the following,

    // optional .android.service.NetworkStatsServiceDumpProto netstats = 3001;
    public static final long NETSTATS = 0x0000010b00000bb9L;

Furthermore, on Android side, this is not strictly proto dependency,
and there's no one using this section, so it can be removed and mark
it as reserved.

Bug: 222625935
Test: FrameworksNetTests
      atest CtsIncidentHostTestCases:IncidentdTest
      adb shell dumpsys netstats
Change-Id: I3083a43a2f9c20768264ce2a531e0a0593f31697