[Motion] Split-shade expansion: implement over scroll effect [v1]
Before, there was code to add over scroll on shade expansion, but only for the notification side of the panel.
This CL adds specific split-shade logic for over scroll, which adds over scroll to the entire panel.

Note that this first version still doesn't take into account fling velocity for the overscroll amount.
Also, this isn't a regression from the current state.

Test: Manually
Test: NotificationStackScrollLayoutTest.java
Test: NotificationPanelViewControllerTest.java
Test: ShadeTransitionControllerTest.kt
Test: SplitShadeOverScrollerTest.kt
Fixes: 228022047
Change-Id: Ic6de84d831a4a6e65139c1cdc04a2612b72dcb49