Invoke BrightnessSetting callbacks even with identical brightness.
Even if the brightness is the same we still need to invoke any listeners
as just the act of setting the brightness has side effects, like
clearing the temporary brightness. Skipping these callbacks mean we can
possibly get stuck in a state where we don't adjust the brightness at
all until the next time a user touches the brightness slider -- this
persists even through screen off.

Fixes: 218600709
Test: Manual
Change-Id: I918636a61424a51f9d22d32fd692beb6eb2b8f2d