Split-shade: prevent new shade expansion when shade already expanded
When not in split shade, we can be on QQS and then fully expand the
shade by swiping down on QQS.

When on split-shade, that logic is still there and causes the shade
to try to expand again even though it is already fully expanded.
This then puts the shade in a weird expanding state even though it is
already expanded and interacting with the shade/notifications doesn't
work.

Test: Manually
Fixes: 229829830
Change-Id: Ia2d3de4d2d09cff96f2621544ce3f9b73a7d36d3