Remove lowpan internal API
This API is not used: marked as @hide with a commented out @SystemApi;
the AIDL server is not implemented in AOSP nor internally. There is one
extra client from the SEPolicy: wpantund; but there is no build rule for
this binary, and no reference for it in AOSP. This API appears be an
artifact from the initial import of Android Things back in 2017.

I confirmed with the previous owners that there is no intention to use
this API in the future (see bug).

This changes reduces the size of framework.jar by 77kb on
bramble-userdebug.

Bug: 235018188
Test: TH
Change-Id: Id66f5ed23ed26c7185f36454f168e34997be3f8f