Initial work profile screenshot policy handling
When a request is received, screen contents are inspected
to determine the most appropriate owner. If the primary
content belongs to a work profile app, the screenshot will
be isolated to the task content.

See: go/work-profile-screenshots for details.

This change also enables the foreground app package name to
be provided with uievent logging.

Additional functionality to follow.

To activate, enable the flag WORK_PROFILE_SCREENSHOT_POLICY

  adb shell cmd statusbar flag 1301 on

This also requires the flag SCREENSHOTS_REQUEST_PROCESSOR

  adb shell cmd statusbar flag 1300 on

Bug: 159422805
Bug: 231957192
Bug: 200294244
Test: atest RequestProcessorTest
      manual; enable flags,
      take screenshot of launcher
      take screenshot of work profile app
      observe logs and screenshot content
Change-Id: Idc74b04ddd449bfd9765ab9a456e9f9344c3aa2c