Track falsing touches on UserSwitcherActivity
UserSwitcherActivity uses the FalsingManager, but the FalsingManager
wasn't being told about touches happening on its surface. Therefore,
the FalsingManager was relying on whatever the most recent touch
was elsewhere.

If this touch was a swipe (such as on the shade) it would not let an
owner change users because a swipe is not a tap.

Test: manual
Fixes: 238598985
Change-Id: Ic1f86d6f4b55b5a28d7db5d56afe72136dbdcef9