Skip syncing local window for rotation with shell transition
The RenderThread might be blocked by waiting for free buffer.
Then it will also block the ui thread of "Pointer location",
which is "android.ui". And then watchdog will restart the
system because the important thread is no response.

This is a temporal workaround to unblock test. It is rare that
system shows a window. So this doesn't affect common use cases.

Bug: 234585256
Test: adb shell setprop persist.wm.debug.shell_transit 1; reboot
      Enable "Pointer location" in developer options.
      Rotate display many times and no watchdog timeout.

Change-Id: I414a2c670aebb78a0418f550015e420eed5f25f4