Make extra navigation bar count towards non decor frame (3/3)
This is the final step to actually make task bar as navigation bars.
This requires all automated test moved to a newer version of UI
automator to avoid missing the area of task bar. Without that, dozens of
tests will fail because the UI Automator cannot correctly click in the
area of task bar.

Also, this partially reverted commit
5148ae9aeb9c5b3bfae6865de1c4cde1e10337ac.

Test: PlatformScenarioTests
Test: TaplTestsQuickstep, ThemeIconsTest, TaplTestsLauncher3
Bug: 238981445
Bug: 238985243
Bug: 238581838
Bug: 233945217
Change-Id: Ibdd343418eeabebdae09153beeaffb975264ad43