Support display-rotation during a split-to-recents transition
Some test automation will dismiss split-screen to launcher and then
immediately rotate the display. Currently, we can't properly
transactionalize the display-rotation, so it just gets squashed into
the existing collecting transition. The result is that the split-
screen handler expected just a split-screen transition but also got
a display-rotation.

This CL adds a mixed-handler case for rotation+split-screen that
stage-coordinator can then use to divide the transition into two
animations.

Bug: 244486556
Test: atest NexusLauncherTests:TaplTestsTaskbar#testLaunchAppInSplitscreen_FromTaskbarAllApps
Change-Id: I82b6a20155dc84e58ab95964da64c9c741f761d9