Fixed camera streaming not being blocked when user taps camera on personal streaming apps
Root cause: When accessing cross profiles using personal + corp account
login, etc., the virtual device always use the user id of the work
profile, whether the user launches the personal profile app or the work
profile app in app streaming. It will cause a NameNotFoundException to
occur and cannot block the camera access when using
PackageManager.getApplicationInfo() to query apps that are only
installed on the personal profile.

Solution: Get all the existing user ids to find out the correct uid of
the app that opened the camera, and compare it with the app detected on
the virtual display to decide whether to block the camera access.

Bug: 245250121
Test: Manual
Change-Id: Iea062154ea472d589e5ffc33998698b4b685d7bd