Remove the restricted user enforcement in getAppExclusionList
This change allows retrieval of the app exclusion list for
callers with the SYSTEM_SETTINGS and INTERACT_ACROSS_USERS_FULL
permissions, as well as allowing UID updates internally when
packages are added/removed.

In both cases, the retrieval of this information does not change
the configuration of the system in any incorrect way, and not
allowing these two use cases results in system components
(Package change notifications and settings app) calls failing.

Additionally, in restricted users, neither provisioning a VPN
profile nor adding app bypass rules are supported, so this
codepath is effectively a no-op at the moment. In the long term,
if VpnManager VPNs are supported from a DevicePolicyManager
standpoint, this correctly updates the restriction list when
apps are added/removed.

Bug: 247344442
Test: atest FrameworksNetTests
Test: atest CtsMediaHostTestCases
Change-Id: I9e820f89a274ea9e26e95ce0aaa6807175760186