Adds back using WindowState for magnification letterbox bounds.
This is a partial revert of
commit c83d192dadab578a64c90c5133c5ac59c5c15a45.
Magnification bounds still use WindowStates for its window iteration
instead of InputWindowHandles, so that change shouldn't have fully
removed WindowState letterbox bounds handling yet.

This change also includes a fix for the bounds calculation: the previous
logic assumed that the letterbox insets were negative instead of
positive. This new version reverses that assumption.

Example bounds for an app in landscape mode, letterboxed to the center
of the screen:
      app bounds       = Rect(663, 0 - 1677, 1080)
      letterbox bounds = SkRegion((0,0,663,1080)(1677,0,2340,1080))

Bug: 248059823
Test: use magnification on a letterboxed app
Test: log app & letterbox bounds, ensure they match expected values
Change-Id: Ic77474c416d0e7097428ea0303302958701e8387