Introduce Expandable to launch activities from Composables (1/2)
This CL introduces an Expandable {} composable that can be used to
easily animate an expandable element into an Activity. This Expandable
will be extended in ag/20139844 to also support dialog launches.

See b/230830644#comment11 for videos and ag/20139924 for an example.

Bug: 230830644
Test: Manual
Change-Id: I1ab9ee7052c36418b1c22155bd2281a84454ec16
Merged-In: I1ab9ee7052c36418b1c22155bd2281a84454ec16