Migrate clock and lock icon to flows
Use transition repository to begin replacing doze amount
listeners. Since a lot of this code will be in Java, provide an
adapter to collect this flows.

For testing with Java code, create two test files. The Java one that
tests with the flag turned off, and a new Kotlin one with the flag
turned on, in order to test with coroutines. Create a base class in
order for both to work.

Use flags to test in teamfood first.

Test: atest LockIconViewControllerTest ClockEventControllerTest
LockIconViewControllerWithCoroutinesTest
Bug: 195430376

Change-Id: Ia5d440e1a6d7c4f526aa508c89e04b4a01abe35e
(cherry picked from commit 83d37b310f50a98e83b1043da09d038de5e84c5f)