Deliver handful of system broadcasts as unordered.
Several existing broadcasts are being sent as ordered simply to
receive a "resultTo" signal when dispatch is finished.  This was
reasonable in the legacy stack, since only one receiver was
dispatched at a time, but the modern stack can dispatch to multiple
receivers in parallel.

This change updates BOOT, SCREEN, and PACKAGE broadcasts to pivot
to this new approach when the modern queue is enabled; we retain
the previous ordered behavior when the legacy queue is being used.

Bug: 253226131
Test: atest FrameworksMockingServicesTests:BroadcastRecordTest
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Test: atest FrameworksMockingServicesTests:BroadcastQueueModernImplTest
Change-Id: I81299b9ec6f509f5b73dfe9484aca12794eb3868