Content provider for quick affordances.
This is the content provider that wallpaper picker will be using to
query for slots, affordances, and selections, and for setting or
unsetting selections.

The CL also includes a contract definition and "client" utility
functions that any customer can use to query and update the content
provider.

Fix: 254857637
Test: end-to-end unit tests included. Manually verified the API by
issuing adb shell content commands and seeing affordances appear and
disappear on the actual lock screen.

Change-Id: I08eefc931dcc7b5133c412f7b97715fedfe84f30