Defines config for default affordances.
As per a request from the large screen team, adding a way for device
configurations to be able to supply System UI with default affordances
for slot IDs.

The way it works is:
1. An OEM overrides the value of config_keyguardQuickAffordanceDefaults
   in config.xml with a list of key-value pairs. Each pair is a slot ID
   followed by one or more affordance IDs.
2. The KeyguardQuickAffordanceSelectionManager contains logic to default
   to those affordances in cases when the user hasn't yet selected
   anything for that slot. Once the user selects something for the slot,
   even the "None" option (which clears out the slot), the default is
   ignored.

Fix: 257051288
Test: manually verified by temporarily adding
<item>bottom_end:home,wallet</item> to the config.xml and seeing that,
if I cleared data for system UI and restarted it, the home control
affordance was visible (after I set it up through quick settings home
tile, for course - because clearing data forgets that too). Also added
unit tests.

Change-Id: I63484a92cd8ed8bbcb6a331f500531c62a021f6c