Adds RearDisplay education dialog to SysUI
To accompany the rear display API's to allow
apps to move their content to the cover screen
on a foldable device, to use the powerful rear
cameras for selfies, there needs to be a consisten
educational dialog to let users know what is going
to happen. This will be shown for all apps that use
the rear display jetpack API's, to keep consistency
and remove any knowledge an app would have to know
about the device they are on.

Bug: 207686851
Test: Manual RearDisplay flow testing
Change-Id: Ib9e2b286c0d84e2411f19f7cf0b0397486e4d4b1