DO NOT MERGE CameraManager: Provide flag for overriding camera output to portrait.
Apps commonly do not handle landscape orientation cameras correctly. In
order to prevent stretching and rotation issues in these apps, this
patch adds a flag to override the behavior of these landscape cameras
to produce a portrait image instead by changing the SENSOR_ORIENTATION
reported by CameraCharacteristics and applying a 90 degree rotate and
crop. It also adds a CompatChanges entry which will track which apps
require this treatment.

Test: Ran on foldable device with several camera apps to verify behavior.
Bug: 250678880
Change-Id: I92b6f227916b97828ab6a7f856eea5a1e427644d