Allow international characters input with apostrophe and double quotation
It has been reported that the behaviour of Android in US international
layout is not the same as Chromebook; according to Wikipedia, the
apostrophone and double quotation mark are "dead characters" that can
be used to enter characters with acute accents and umlauts:
https://en.wikipedia.org/wiki/QWERTY#/media/File:KB_US-International.svg

Since this seems to be the behavior in Mac and ChromeOS, this change
will align Android with the rest of the platforms. I decided to follow
the same pattern as MacOS, since typing the dead character twice will
input the actual character, and apostrophe/quotation mark are very
useful to have.

Bug: 259862772
Test: confirmed on device that it works as expected
Change-Id: Iaf326e07e9b07d07383be73ed65b1cac338ae0bf