Check lock-task mode if task_on_home is applied
f6202e18dcb0df47c, always checks whether the home task is violated
with lock-task mode while moving the task to the front. That result in
the pinned toast was shown even applications don’t not request to have
task_on_home behavior.

This CL checks the home task violation with lock-task mode after the
task-on-home flag is applyed.

Bug: 261383209
Test: atest TaskTests#testReturnsToHomeRootTask
      atest ActivityVisibilityTests#testLaunchTaskOnHome
Change-Id: Iccc7e595b043a712c5959d28ba14f54825bcadd6