Set orientation for boot animation and default display rotation.
Change the default display rotation value and boot animation orientation to the value specified in the sysprop ro.bootanim.set_orientation_<display_id>. Four values are supported: ORIENTATION_0, ORIENTATION_90, ORIENTATION_180 and ORIENTATION_270. If the value isn't specified or is ORIENTATION_0, nothing will be changed.

This is needed to support having default rotation behavior and boot animation in orientation different from the natural device orientation. For example, on tablets that may want to keep natural orientation
portrait for applications compatibility but have landscape orientation as a default choice from the UX perspective.

Bug: 260627934
Test: manual test with all 4 values and atest WmTests still pass
Change-Id: I322919999ef787ff2c149293d7e1c7e669796c9a