Add method in WallpaperManager to get config_independentLockscreenLiveWallpaper
LockscreenWallpaper.java (and potentially other files) have access to a WallpaperManager instance, but no instance of context/resources.

To minimise changes, a hidden method boolean isLockscreenLiveWallpaperEnabled() has been added to WallpaperManager to get the config flag.

Other classes can use this flag and implement the new wallpaper logic when it is true.

Bug: 197814683
Test: treehugger (low-risk, no logic changes)
Change-Id: I791ba72f571f996906d2c23f97bba7b88b2fe21f