Refactor TaskFragment related WindowContainerTransaction
Refactor TaskFragment related WCT to a single transaction type. As we
will likely to add more TaskFragment transactions in the future, putting
them into a standalone type can make them more organized, as well as not
confusing other users of WCT, especially because WCT already have a lot
of operation types.

Refactor deleteTaskFragment to take fragment token instead of
WindowContainerToken to be consistent with other TaskFragmentOperation.

Remove support of setAdjacentRoots/clearAdjacentRoots for
TaskFragmentOrganizer since we have setAdjacentTaskFragments.

Remove reparentChildren because it is never used.

Bug: 263436063
Test: pass existing tests.
Change-Id: If01fde450f03f1f6677e583f4bdf97fcd1192553