Setting the display state to off when the default display is being changed
This fixes the issue of the display state change always being "in progress". PhotonicModulator thread fails to make any progress while waiting for a notify call to release the lock. This happensbecause the screenState is set to invalid when a device is being folded/unfolded, forcing it to wait until the screen state is recovered.

Bug: 262294651
Test:  adb shell dumpsys display | grep "Photonic Modulator" -A 10 ->
Shows mStateChangeInProgress is false
Test: Manually folding/unfolding the device wont cause the display stuck
in the power off state

Change-Id: I903f6392cbe42031aab8ffdbfbe541463d6b5e01