Revert "Enable tethering namespace accessible in Android native level."
This reverts commit f177a298113a36e8a919f2cb6bb21c48aeded1c9.

Reason for revert: Original commit was for native experiment setup.
But SEPolicy was not updated and sync to native level always fail in T devices.
This CL will clean up incomplete native experiment setup.
Even if we decide that we want to use flags from native code in the Tethering module,
we will do so using a new namespace, tethering_u_plus_native.

Test: m
Bug: 264947218
Change-Id: If32fe0b237dcb213886efb72430ea7575ab64530