Attributing and logging wifi wakeups
If Wifi is included in the transports of the network the packet came
on, we classify the wakeup as a Wifi wakeup.

Currently, only the packets arriving on the wifi interface are marked,
so we only do this for Wifi. We can add more subsystems once more
interfaces start marking these packets in the lower stack.

Test: Manually inspect the output of `dumpsys batterystats --wakeups`
and `statsd_testdrive 588`

Bug: 265742148
Change-Id: Id444757330bf776befd78116194e05e78a00f3ad