Reset title message in keyguard when onThemeChanged is called.
If keyguard is visible when wallpaper colors changed event is
propogated, title message of the keyguard was disappearing. This was
a flaky janky behaviour since those events were not synced, but it
will be consistently visible after fixing b/266711330.

Bug: 266940281
Bug: 266711330
Test: Tested manually by delaying user switch events (via ag/21145270)
to make the jank consistent, and made sure it's gone after the fix is
applied.
Test: atest KeyguardSecurityContainerControllerTest
Test: atest KeyguardPinBasedInputViewControllerTest
Test: atest KeyguardPasswordViewControllerTest
Test: atest FrameworksServicesTests:UserControllerTest

Change-Id: Ifce985363eee20ab195eb6b2fdfc6dcf175ad9f1