Long-press not required for mouse or stylus.
When interacting with lock screen shortcuts, a long-press is required,
mainly to help protect against false clicks (e.g. "falsing"). When using
a stylus pen or a mouse, it is no longer a requirement, as those pointer types offer a  much
greater degree of confidence that the click was not a false click.

This CL takes care of it by checking the type of pointer and, if it's a
stylus or a mouse, allowing a simple click gesture to happen, instead of the
complicated long-press.

Fix: 267534467
Test: manually tested with a stylus - simple "launch" affordances like
camera work and require a single tap. Complex "toggle" affordances like
do not disturb toggle on and off, requiring a single tap
Test: manually tested without a stylus. "Launch" and "toggle"
affordances still require a long-press, the animation still works. When
tapping them with a short click, the shake animation/vibration and
educational text still appear as expected.

Change-Id: I92270e554f41b0066f5e8c8fec52c5d7c6c93b29