Add flag for disabling transitions from ARC to eARC
This flag is introduced to disable transitions from ARC to eARC if we
see any issues with the current design.

If the flag is enabled, the behavior is as in the original eARC design.

If the flag is disabled, eARC now gets disabled in the HAL when ARC gets
enabled in the HAL. As a consequence, the HAL won´t attempt to establish
an eARC connection if ARC is already connected. eARC will be re-enabled
in the HAL on 2 occasions:
- on the next wakeup or boot
- when the eARC setting on the TV  gets toggled to 'enabled´.

Test: make and run unit tests with the flag set to false
Bug: 264414108
Change-Id: I77037718e24c7b14532cb6d684c053d28c222659