Touch handling and other tiny cleanups in NPVC
The interface used for the touch handling handoff between Status Bar
and Shade is nearly the same as Gefingerpoken, so this change will
remove the redundancy. It also removes the usage of the view as an
an unnecessary middleman by passing events directly to the Shade
touch handler. Also includes some trivial warning cleanup. This was
originally the first in a chain of CLs attempting to fix multitouch
in the shade, but the follow up CL was too risky for QPR.

Bug: 132458655
Test: manual and atest
Change-Id: Iabb3484e43573b6e4f7755d1abe0a76512828834