Allowlist permissions in a better way.
All apps should be UPGRADE_EXEMPT by default, which will be replaced
by INSTALLER_EXEMPT when onPackageInstalled() is called and the
installer supplies the list of permissions to be allowlisted (most of
the times just the default of all requested permissions).

This way we can properly handle the system app case and the rollback
case, in which cases there won't be any installer to allowlist
permissions.

Bug: 263504888
Test: RestrictedPermissionsTest
Change-Id: I58d6d7aeb8db8de05c2e6ac60b51358709f06091