Avoid NPE when removing fixed rotation transform
The legacy code "checks c.getParent() != null" in resetTransform
to avoid the case such as transition finished after the token is
detached.

While with shell transition, the rotation leash still needs to
be removed but ignore reparent in case the token is detached.

Also extract FixedRotationTransformStateLegacy for legacy
transition. So by default the shell transition can skip the
unnecessary creation of SeamlessRotator.

Bug: 277697318
Test: IncompleteMotionTest
Change-Id: I6e5ebd15a972d88232e7a9a46ff1ab4f557c58d7