Handle uiMode change in CompatUI
When uiMode changes we need to recreate CompatUI components to
allow the update of the resources in case of switch between
Light/Dark more or when the device is docked/undocked.
This allows to cover the use case of a change in theme of
a docked device showing a restart dialog.

Fixes: 279170801
Test: atest WMShellUnitTests:CompatUIWindowManagerTest
      atest WMShellUnitTests:ReachabilityEduWindowManagerTest
      atest WMShellUnitTests:RestartDialogWindowManagerTest

Change-Id: I179df47bdc23f238dee9ffe6b2e579a67921eb1f