Delay boot complete broadcast for main user on HSUM.
This fixes an issue where on a HSUM device, the LOCKED_BOOT_COMPLETE
broadcast occurs before boot actually completes.

We already delay the sending of this broadcast for user 0, but we
need to do the same for the main user because otherwise the
broadcast is sent at the end of user switch, and the switch to the
main user occurs before boot complete.

Bug: 279983151
Test: atest UserControllerTest
Test: build & flash; observe logs during boot
Change-Id: I885954ae57b3c7bc27b2f03b2d2bfa8583c79293