Fix a transition issue when entering a dream.
When entering a dream, the transition should be for the visible activity
to remain in place while the dream fades in. A bug was causing the
current activity to slide out before the dream faded in.

Test: Manually by launching setting up the device to show a screen
saver, launching an activity (such as Play Store), and then waiting for
the device to show the screen saver. The previous activity should remain
in place as screen saver fades in.
Bug: 279487452

Change-Id: Idccd30cdf26fca9826a80d9705f17217d92034a7