[SB Refactor] Only use carrier merged sub ID if wifi or CM are default.
(or the device is in airplane mode)

The old pipeline's WifiSignalController only sent updates to
MobileSignalController if the wifi info was carrier merged *and* wifi or
carrier merged was the default connection or the device was in airplane
mode. The new pipeline previously didn't check this default status,
which caused the new pipeline to keep polling the carrier merged info
instead of the normal cellular info.

Also increases the size of the connectivity buffer, since investigating
the BRs attached to this bug revealed that the buffer cut off too early.

Bug: 278618530
Test: atest MobileConnectionsRepositoryTest
Change-Id: Ib9dde0234c38921bb67179a138f9cc8c60e7d669