Add a config_brightWhenDozing to suppress the dim state when we entered dozing wakefulness but haven't started dream service yet.
Normally when we enters dozing wakefulness state, the power policy will
be set to POLICY_DIM for a short period of time before dream service
started. This config will suppress this effect and force the policy to
be set to POLICY_BRIGHT and completely get rid of the dim temporary
state.

This config is expected to be used when config_maximumScreenDimDuration
is set to 0.

Bug: 280324422
Test: manual test AOD on and AOD off case. Dim state is skipped.
Test: atest PowerGroupTest
Change-Id: I5f060e46389dee70ce945eb2e84f45d5927ca606
(cherry picked from commit 971a3e143019b1ee93881f87bd75ee8c00a226f8)