Hide fullscreen button in partial only mode
The fullscreen button is hidden when the capability of magnification is
set to magnifying partial screen only. The logic for updating UI control
is to check the capbility of magnification first to decide if we would
like to show or hide the fullscreen button. If the capability is
ACCESSIBILITY_MAGNIFICATION_MODE_ALL, we check the current mode in the next step to
decide if we would like to show or hide the edit button and the diagonal scrolling view.
We also register a content observer for modifying the layout immediately
when there is a capability or mode change in magnification.

Bug: 280520144
Test: Manually - attach video with the bug
Test: atest WindowMagnificationSettingsTest WindowMagnificationTest MagnificationSettingsControllerTest
Change-Id: Ie43fdbc5bf598130097deb3db718f0c1535afbb4