Use empty string instead of nullable for leak report email
Leak report email was set to be nullable in some modules and not
nullable in others. This was causing undeterministic build failures when
building SystemUI with gradle.
Removing nullable annotation from leak report email dependency and using
empty string instead of null.
Only setting the email in leak report email intent when the email string
is not null or empty.

Bug: 281537214
Test: full build: m -j

Change-Id: Ic798ae8aba76ba29f9ded8dd8a549fd4c31962a6