Open intent with animation for mini-resolver
Uses an animation when the min-resolver opens
an intent (push down and fade out the mini-resolver,
then sliding in the new activity).

Note that this does not work if the new activity is opened
in a new task, due to being unable to override the system
animations for task transitions.

Bug: 271818896
Test: Manual, see bug for before and after video
Test: atest IntentForwarderActivityTest
Change-Id: If64da935621f49e5c5013dd791b85f48c0cfd861