Handle openning transition while animating a pair to home
There is a timing opening one of the splitting apps while animating a
pair to home. Which should launch the app into fullscreen instead of
bring back the split pair. This makes sure to dismiss split properly
with the openning transition. Also makes sure to restore reparent leaf
task flag to prevent split screen being dismissed by relaunching a
splitting app while it is in the foreground.

Fix: 280387453
Fix: 283462648
Test: atest WMShellUnitTests
Test: compose a split piar, swipe-to-home and quickly click one of the
      app icons, it won't bring back the split pair
Change-Id: Ia4e0b5f74f3b980ab3ecce18bebb14fcb9530f87