Add caption inset when moving to desktop
When we transition to desktop using the caption handle, the
configuration changes trigger WindowDecoration#relayout to run and a
redraw. WindowDecoration#relayout adds the caption insets source using
another WindowContainerTransaction, but, the insets are sometimes not
added before the redraw occurs and apps do not listen to inset changes
to trigger a redraw, so, the insets are not added until another
configuration change triggers a redraw.

In the long run, we need to have apps listen for inset changes and
redraw accordingly. In the meantime, this change adds the caption insets in the same WCT that applies to changes to transition into desktop mode.

Test: Repeatedly transition into desktop mode using the caption handle
and make sure the caption inset is there (top of app is visible) + run
atest DesktopTasksControllerTest

Change-Id: I16260aa21374fee751f03fe4492f91640179a7ad