Pass LayoutParams from SCVH to Parent window
There are some params that a view or window can set that needs to be
handled by WMS. However, SCVH have no interface into WMS since they are
handled in the client process. Instead, pass the params to their host
parent and let the host ensure the correct params are handled in WMS.

Test: SurfaceControlViewHostTests
Bug: 281687307

This reverts commit 38b4f00a28a4ce12850f765541e4fed80880a3c6.

Reason for revert: Disable failing test. b/286245842

Change-Id: If0e61bfc4d1e32513316d4a9a22faf2ec77e690c