Fix TV PackageInstaller AlertDialog minimum button bar height
AlertDialogLayout gives the buttonPanel its requested minimum height,
or that of the buttonPanel's single child.
The buttonPanel is a ScrollView with padding that contains a single
ButtonBarLayout. The ButtonBarLayout calculates a minimum height of its
button's height + its padding.

But since the ButtonBarLayout itself doesn't have padding, its minHeight
is the exact height of its buttons. The AlertDialogLayout lays out the
buttonPanel ScrollView with that height, but the ScrollView applies its
padding, pushing the buttons out of bounds and clipping them.

This change moves the padding to the ButtonBarLayout, so that it takes
the padding into account when calculating its minimum height.

Bug: 280552754
Test: m
Change-Id: Iafbf9b027a399c05c63f3123e5e924ec50584dcf