Fix StemKeyGestureTests
StemKeyGestureTests was failing after ag/23249820, where the stem key
behaviors are fetched in PhoneWindowManager#updateSettings() (instead of
PhoneWindowManager#init()). Since updateSettings() was mocked out in
this test, the key behavior was not being read properly.

We now enable the updateSettings() method in PhoneWindowManager for
tests, and use the Setting values of the key press behaviors to
configure test behaviors.

Bug: 287995215
Test: atest WmTests:StemKeyGestureTests WmTests:PowerKeyGestureTests WmTests:SingleKeyGestureTests WmTests:ModifierShortcutTests WmTests:KeyCombinationManagerTests WmTests:CombinationKeyTests WmTests:PhoneWindowManagerTests
Change-Id: I9ef0ab5923e124b350497b2f176b34e4f275a47d