Fix letterbox overlapping with status bar icons after rotating device
When the device rotates or status bar icon bounds change, we re-use the
previously cached letterbox details that were sent from WindowManager.

Also, when calculating whether there was an overlap between the icons
and the letterbox, we were using the Rect#intersect API instead of
Rect#instersects, which actually mutates the Rect, which was then
cached, and used for the next calculation.

The fix is to use the API that doesn't mutate the Rect.

Fixes: 287508741
Test: LetterboxAppearanceCalculatorTest.kt
Change-Id: I3c6be62ef673b674f3305ae254494dd41c5cce29
(cherry picked from commit 38f66e1ffb2aa72e66713093f3e14eb82e28975e)