Adds a method to TaskView to allow clients to set caption insets
Some users of TaskView (bubbles) need to be able to add a caption
bar to TaskView to allow users to manipulate the bubble window or
access a menu.

This CL exposes a method on TaskView to set a rect that represents
the insets for that caption bar. This rect is automatically excluded
as a touchable area so that TaskView doesn't steal touches from it.

Test: manual - locally use new setCaptionInsets method on a TaskView
             => observe that content is inset & touchable
Test: atest TaskViewTest
Bug: 284205099
Change-Id: Ia71f56920bb85b6e0e7164c92273caf4d480916a