Reconnect if weaver service connection dies
For robustness reasons, reconnect to the Weaver service if it dies.
This should not be relied on, but it's easy to do and is the same thing
that LockSettingsService does for the Gatekeeper service.

Bug: 287705522
Test: Treehugger
Merged-In: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
Change-Id: Ic06c45353c6b7282e2f22eef13ea1c6e36fe5c58
(cherry picked from commit 9fc93b1ba75bcf6ef0229c44d7a49eacfeae0d7c)