Revert "Avoid changing z-order of overlay layer after resizing display"
This reverts commit 05c4922e060ae029d0e72823bd61b2bbd03f34f7.

Reason for revert: b/288726181

The effect of
SurfaceControl.Builder#setParent(null);
and
SurfaceControl.Transaction#reparent(sc, null);
are different.

The former is to put at root layer, but the latter is like
to detach the layer. So the flickering was disappeared by
not showing the layer. That is unexpected. Currently there
seems to be no way to assign root layer by transaction.

Change-Id: I5b903d5dc7c30b644365b4b28048e74453e45922