Add a failsafe to update surface position after activity-level transit
ActivityRecord isn't expected to change independently, however, there
is now compat behavior that will enable client-code to trigger changes
to an AR's surface outside of any WM-side lifecycles.

This already causes glitchy offsets w/ legacy during animation, but
w/ shell, the activity could get stuck with the offset if the compat
properties were changed during the animation.

Until we come-up with a proper solution for this, just "sync up" the
surface state when idle so at-least it won't be stuck. It will still
have an offset during the animation, but that is the behavior we
already had.

Bug: 282129721
Test: Use a forced-landscape device (eg. unfolded landscape) and
      launch an activity which has portrait rotation in manifest
      but then requests landscape rotation in onResume before
      animation finishes.
Change-Id: I92ee7884687433b12bc2649ade7416abb9953943