Disabling the Done button when system applying new font size
Since we are not able to interact with the whole screen when the system is applying the new font size, we disable the Done button before updating the font scale and enable it again after the change finishes. The Done button is turned into gray when disabled to let users understand that we are not able to interact with the button at that time.

Bug: 288699753
Test: atest FontScalingDialogTest
Test: manually - attach video to the bug. Manually check that the button is disabled when the system is updating font size and the button is enabled again after the font size change finishes.
Change-Id: Id56bb41290e44ea922031f3515c484ef7dc1c1b5