Add additional arguments to PackagesHash.
RetailDemo role depends on additional states of the device. Adding those to the package hash as well.

Security approved for design: go/retail-demo-role-fr

This change is a re-do of ag/23720967, DevicePolicyManagerInternal could be null on some form factors. Since this is in system server, we won't check the feature present but do null check directly.

Bug: 274132354
Change-Id: Ia8b1b1fbac06df19f93f85929fc8178a5a05fd9c
Test: manual