Remove NetInitiatedAcitivity in AndroidManifest
This should have been done in ag/24302901.

Please see the rationale in b/294306993. NetInitiatedActivity is a rarely used code since we introduced the GnssVisibilityControl HAL to replace GnssNi in Android 10. The earliest android version that can upgrade to Android U is 10 so it's okay to delete the GnssNi related code.

Bug: 294306993
Test: on device
Change-Id: I53c46d28d74df15fba71d4431d83429759ab845c