Adding performHapticFeedback to ShadeViewController.
The ShadeViewController interface is expanded to include a new method
to deliver haptics using the performHapticFeedback method from a view.
The method should be called using the provided constant passed as a
parameter. This is used in the CentralSurfacesCommandQueueCallbacks when
a vibration is required on the KEYCODE_SYSTEM_NAVIGATION_DOWN event. The
changed is controlled by a flag.

Test: atest CentralSurfacesCommandQueueCallbacksTest
Bug: 245528624
Change-Id: I22875afd5537d6dd149c8ec9d0619617dae95bff