Make the TV start window exit animation a fade out
Prior to this CL the default animation for the start window exit
animation was always a radial vanish+slide up. On TV we want to
change that to a fade out animation.

This CL introduces a config to specify the type of the animation.
We add a fade out animation as an alternative to the radial vanish+slide
up exit animation.

Bug: 296354973
Test: manual - check that changing the config alters the exit animation
      to fade
Change-Id: I62f3a4deadb85c4839321dc1543122825cb98bb7