Validate underlying network when VPN switches to unvalidated
Validation failures on VPN networks may be caused by problems
on the VPN network itself or a broken underlying network.
However, the existing mechanism can only detect if the VPN
network is usable or not.

To clarify the cause, this commit adds a mechanism to trigger
underlying network validation when VPN receives validation
failure. This will not only help to clarify the connectivity
issue but also trigger the possible system default network
switch if the breakage happens on the underlying network.
It might also trigger recovery (in the case of cell).

Bug: 296183603
Test: atest FrameworksNetTests
Change-Id: I4a2964001e0787b2c198fdcb2ef9d0819a45b4e8