Pause activity before resuming the next one
The occluded activity was still visible while starting an Activity
on a new TaskFragment on top because the occluded TaskFragment was
still adjacent to the pinned TaskFragment. The state was unset until
the TFOrganizer gets the #onTaskFragmentAppeared callback and perform
updates via WCT.

This CL makes sure the TaskFragments are updated along with the
activity launch. So, the activities on the occluded TaskFragments
should immediately become invisible.

And in the WM Core, similar to commit 7b5a574, we should pause all
non-top child TFs in the Task if needed.

Bug: 300019674
Test: atest PinActivityStackTests
Change-Id: I7505ffe4b9d664fe8858b83c3744324401ed3db6