Refactor pending display logic
Introduces the concept of ignored displays in DisplayRepository, and handles the case of more than one pending display.

The PendingDisplay interface is now both in DisplayRepository and ConnectedDisplayInteractor as we the data layer should only communicate with the domain one, and the domain one with ui one, but Interactor pending displays are essentially just proxies.

The new `ignore()` method removes the pending display from the repository. `disable` has been removed from ConnectedDisplayInteractor as not used for now.

+ use isKeyguardShowing instead of isKeyguardUnlocked. Before, we were showing the dialog also in the lockscreen if the unlock method was just swipe. Now, just having the keyguard visible is enough to prevent the dialog.

Test: MirroringConfirmationDialogTest, ConnectedDisplayInteractorTest, DisplayRepositoryTest
Bug: 298023961
Change-Id: I06477a9aa652696d8db5807414ee708fb4fe41f9