Hide caption when task is in immersive mode
Adds DesktopModeOnInsetsChangedListener which listens for inset changes
and checks to see if the status bar inset has disappeared, signifying
that the task is in immersive mode. If so, the caption will become
invisible. When the status bar inset reappears, we know we have exited
immersive mode and make the caption visible again.

Test: Open Youtube and expand window to fullscreen. The window
decoration should disappear. When we exit fullscreen, window decoration
should reappear.
Bug: 286344314

Change-Id: If913a81e94404c21b544d6e6bba6c35ab2906d8d