Add config flag to disable Pinch resize PiP
Currently, `mEnablePinchResize` is controlled with `DeviceConfig`,
which is in my understanding useful for launching the feature,
doing A/B testing etc, but with `DeviceConfig` there's no reliable
way for a vendor to ensure to disable it because it could be
overriden remotely. This CL introduces a new static config value
for this purpose.

Bug: 298139992
Test: Pinch resize is disabled in ARC.
Change-Id: Ic494168d26dea4002a9c52159ef8bd61239a8cdc