Remove scheduleTraversal when applyTransactionOnDraw is called
When a caller invokes applyTransactionOnDraw, it means they want to sync
the transaction with the next frame change. This doesn't mean we need to
schedule the frame change at this time. It doesn't make sense for the
caller to attempt to sync a transaction without making any attempt to
draw a frame.

Also this cleans up code where there's nothing new to draw in
performTraversal, but there was pendingTransaction. The draw code will
return false and the code later on will ensure it applies the
transaction.

Add logs and traces to VRI

Test: AttachedSurfaceControlSyncTest
Test: AttachedSurfaceControlTest
Bug: 302404882
Bug: 302690479
Change-Id: I5add18c672060de924b4efd7e36f442f1ad796cf