Fix the overlapping when a pinned auto-cancel notification is clicked
Fix the problem that when a pinned auto-cancel notification is clicked
when the Shade is opened, and the Shade collapsing is canceled. This
change added the normal dismissal animation for the clicked auto-cancel
HUNs to let them perform the animation correctly when the Shade
collapsing is aborted.
 Moving forward, we should only add transient views:
 1. In the same frame that the view is removed
 2. When the removal animation is currently running, or determined to be run
 We can also consider adding cleaning up procedures that act as
 back-ups for the removal animation end listeners.

 Bug: 297950089
 Bug: 281628358

Test: manual post a HUN with auto-cancel, open the Shade, click the
notif, and cancel the Shade collapsing. There shouldn't be overlapping
HUNs.

Change-Id: I3d5a76c5fc0e1baeb9c8caa4832fb9cce061a51b