Refactor to ecapsulate FadeOutManager configurations
In preperation of supporting configurable fade parameters,
move the default configurations into a separate class.
The fade apis use the getter methods to query the fade
parameters instead of directly accessing the default
static values. This separates the concern of fade parameter
settings from applying them on players. Changes to
configurations in future shall be localized to the new class.

Bug: 186905459
Bug: 302870089
Test: atest -c --rerun-until-failure 1000 AudioFocusTest
        FadeConfigurationsTest FadeOutManagerTest

Change-Id: Id3cc1063d48c0ad5c1935a8ec6603bca8f1720c3