Fix Partial Screenshare always launching split root when launching task from recents
There is a long standing bug with Partial Screenshare where if the target app is part of a split pair and we attempt to launch it from recents, while it is in the background, it will always launch the split root (whichever app started the split) instead of the selected app.

This is due to ActivtiyTaskManager#startActivityFromRecents's logic for launching an activity, instead of launching the task fresh in fullscreen, it leads to StageCoordinator attempting to launch the app in split screen instead of fullscreen.

By leveraging existing logic in RootWindowContainer#getOrCreateRootTask which handles finding the root task to launch, we can instead change the conditional path taken when starting the task, so it will be launched in fullscreen, ensuring that the selected app will be shared instead of the split root.

Bug: 299597333
Fix: 299597333
Test: atest com.android.systemui.mediaprojection.appselector.MediaProjectionAppSelectorControllerTest
Test: atest com.android.systemui.mediaprojection.appselector.data.ShellRecentTaskListProviderTest
Change-Id: I0338956041c750453b781d1e8f407e705f7f9a60