Enable separate keyguard previews based on screen id
The surface view host token is used to distinguish between different
PreviewLifecycleObservers, but the 2 different surface views for the
dual display preview return the same host token. To distinguish the 2
surface views, both a host token and a display id is used as identifier.

This CL is based on changes in the launcher preview renderer:
Iba9518e93c09f0f8304cbce4c3bee9f8753cd583

Test: manually verified
Bug: 303313708
Change-Id: If1a7b9d6b4f7bdbd91b3cdd8f4a9714175e6399e