Display provider icon to inline suggestion
Display provider icon to inline suggestion if the credman provider has
one. If not, use the default drawable that credman bottomsheet uses.

With this change, the inline icon uses the same icon, but doesn't share
the icon color with the bottomsheet. The credman bottomsheet uses color
from Composable, which CredmanAutofillService had trouble importing.
Will look into this.

Bug: 299321128
Test: Build, deployed locally
Change-Id: I3d159ae1ced7d88b7f21eacd5a4a3fb72486178e