Remove the WMShellFlickerTestsPipApps from csuite zip
This test suite requires signature level permissions, which are only
granted to apps signed with platform cert. Csuite is not part of
the platform build and the APK would have not been signed with it.

The platform signed version of the APK can be downloaded at test infra
setup phase instead of relying on it being packed into the same zip.

Bug: 302034872
Change-Id: Ib2c6df280b703ba45e08633852a4e353a68391ae
Test: n/a
Flag: n/a