Invalidate patternview after animation.
In cases where there is a lot of measuring and layouts like the one
described in b/301892345, the pattern view can sometimes appear smaller.
However, when you begin to interact with the pattern view, it reverts to
the correct size. By invalidating the view when the animation is
complete, we can ensure that the view is the correct size every time.

Test: open user switcher, add user, edit name and open keyboard, edit
avatar, and observe pattern view.
Fixes: 301892345
Flag: NONE

Change-Id: Ia562ca7db93fd52febdc9b01bd50c8136062fd88