Initialize DisplayUpdater with the initial display info
DisplayManager#getNonOverrideDisplayInfo might not
set DisplayInfo values if the display doest not exist.
This causes problems when mDisplayInfo in DisplayUpdater
was never updated as it will emit empty DisplayInfo
if we call updateDisplayInfo when the display is
already removed.

The difference with the previous behavior is that
we initialized mDisplayInfo with initial values in
constructor of DisplayContent, so when we request
an update when the display is removed it would use
the old values to perform the update instead of
using 'empty' values.

Bug: 313307263
Test: manual fold/unfold, presubmit
Test: abtd run of CtsDisplayTestCases
Flag: none
Change-Id: I6b09e00b3d7d0fabb45b4df8f747fa5f62b6342d