KeyguardManager: clarify behavior for profile with separate challenge
In my improvement to the KeyguardManager documentation a couple months
ago (https://r.android.com/2747678), I didn't mention anything for
profiles with a separate challenge because it wasn't yet clear what
isDeviceLocked() and isDeviceSecure() return for them and whether they
work as expected in this case.  However they do appear to work as
expected in this case, so update the documentation to cover this case.

Bug: 128589540
Test: N/A
Flag: DOCS_ONLY
Change-Id: I3762a47279f43e2a37c89c49c52380cf08d2f802