TrustManagerService: dump deviceLocked for profiles
Make 'dumpsys trust' show the deviceLocked state for users that cannot
be switched to, including profiles.  Previously, the deviceLocked state
was not shown for such users, presumably under the assumption that it
was always the same as the parent user's.  However, profiles that use a
separate challenge do have their own deviceLocked state.

A related issue which this change addresses too is that the output
contained "managed profile" even for other types of profiles.

Bug: 296464083
Test: Tried 'dumpsys trust' in various situations
Flag: exempt, just changes dumpsys
Change-Id: Ibb949626fd4a7a32ebc6daeeff6e9c0192583009