Add ExecutorContentObserver.
This allows creating a ContentObserver with an Executor instead of a Handler.

This is currently done by creating an awkward constructor on ContentObserver that takes an Executor and an unused int to avoid breaking source compatibility. It is marked with @hide because it is really awkward.
ExecutorContentObserver, which is less weird, is also marked @hide. This could theoretically be unhidden, but will be valuable to platform source even as-is.

Change-Id: Ie3a0f200abcf1834cbf969186dc3a5aa11efb8a4
Fixes: 232409779