No logs if BAL hardened result unchanged
If BAL hardening does not change that we allow the start, there is no
reason to print a warning. Currently this emits a lot of warnings for
cases that are not PendingIntents (regular activity starts). In the
BackgroundActivityLauchTest run this reduces the number of BAL wtf logs
from 169 to only 6.

Test: atest BackgroundActivityLauchTest (and inspecting logs)
Bug: 296478951
Change-Id: I332294eb823c38f02aca34bc18eadfd9facb867b