Add Modifier.noResizeDuringTransitions()
This CL adds a new Modifier to force a layout node to always be the same
size as when there is no transition. This can for instance be used to
make sure that a scrollable container/scene does not resize and thus
that its scrolling state does not change during a transition. This will
for instance be used by the QS pager in the demo app, so that the QS =>
Shade transition is smoother.

Bug: 308961608
Test: SizeTest
Flag: N/A
Change-Id: I98e65d0c1bffb47be04309b743587e957d7ca188