Move scene picker definition inside ElementKey (1/2)
This scenes moves the scene picker definition inside ElementKey instead
of inside a transition. The main reason for this is b/317026105: we need
to be able to easily set the default scene of a MovableElement. Without
this change, we would need to set a scene picker on all transitions
between at least one scene that contains that movable element. For
instance, because the movable MediaPlayer in the demo is in the
Lockscreen, Shade and QuickSettings scenes, we would need to add a
picker for all transitions except Bouncer => Launcher.

Test: Existing STL tests
Bug: 317026105
Flag: N/A
Change-Id: I59953fec3be885d2e787a0705f93b621eb2c4f7d