API to get the sdk sandbox uid for an app uid
There is an existing API, Process#toSdkSandboxUid() which is a system API and we want to make it public. But the naming is not consistent with the Process#getAppUidForSdkSandboxUid() API. To fix this inconsistency, a new API is added.

Creates Process#getSdkSandboxUidForAppUid API which returns the sdk
sandbox process uid for the app uid.

Bug: 318486754
Test: atest ProcessTest
Change-Id: I6dfd2cfbba608c88321d688632196bb4e1e877d1