TelephonyRegistryManager API review changes
In formalizing notifyCallStateChangedForAllSubscriptions and
notifyOutgoingEmergencyCall as proper system APIs in Telephony, API
council has provided additional feedback to expand the docs for the
notifyOutgoingEmergencyCall (who's being notified) as well as backing
the API with a permission requirement.

From other feedback, TelephonyRegistryManager has been promoted as a
SystemService in order to help indicate how to obtain an instance of
this class. The subId field in notifyOutgoingEmergencyCall has also been
expanded to be subscriptionId.

Bug: 321265373
Bug: 321800124
Bug: 311773409
Test: m
Change-Id: I7a192bf11f88d823fa6463550cc6df0d5c78807c