Let view focus move across adjacent task fragments by tab keys
Before this CL, the focus could only be moved to the focusable views in
a window by tab keys if there are multiple focusable views in the
window.

This CL gives the tab key a chance to move the focus to the adjacent
task fragments while the next focus will be looped.

Fix: 283618824
Test: atest ActivityEmbeddingFocusTests
Test: Run commands: `adb shell input keyevent KEYCODE_TAB` while running
      2-pane Settings. See if the focus will go to the adjacent window.
Change-Id: I16c68a514cd7bf9238e2a9fcbc130b025a3ae872