Add input test for a11yInputFilter
This test is about sending input events to AccessibilityInputFilter and
checking what events are sent back to InputDispatcher from the filter
using the IInputFilterHost interface.

Here, we are focusing on the public APIs of the InputFilter.java and
some of the AccessibilityInputFilter. The goal is to treat the
AccessibilityInputFilter as a black box as much as possible, with
minimal invasiveness.

Bug: 310014874
Test: atest AccessibilityEndToEndInputTest
Change-Id: Ia73c270b83d11505dd4fe373649d2033ee6f53e3