Baseline `Trace` support for SysUI.
SysUI code relies on the tracing APIs, which are often used for
investigating performance-sensitive operations on physical devices.

As this tracing information is less useful under the Ravenwood
environment, we offer a baseline no-op implementation to support
existing code.  (There are no APIs to "read back" tracing
information for a test to inspect, which is another reason it's
okay to no-op for now.)

Bug: 319647875
Test: atest FrameworksCoreTestsRavenwood
Change-Id: I52865d7da4c38d350edc7e9dc0239fc1b130048d