Define `SystemUiRavenTests`, default to "ignored."
With our recent foundational work, we're now in a position to define
a `SystemUiRavenTests` target that runs all `multivalentTests`.  This
change uses the `@DisabledOnRavenwood` annotation to "ignore" all
tests by default when executed under the Ravenwood environment, and
future changes will use `@EnabledOnRavenwood` to progressively
opt-in tests that are passing.  (Tests continue running as-is under
all other environments.)

As a reminder, our strategy of using a `RavenwoodRule` and
`RavenwoodClassRule` to integrate with JUnit means that all test
classes <clinit> will be run, even if all contained tests will end
up being ignored.  We've done a few passes of using `by lazy` and
`lateinit` to ensure that we avoid interacting with Ravenwood
unsupported APIs, but once this change lands it will require that
future test updates remain vigilant to keep presubmit green.  This
aligns with the SysUI testing policy outlined in <go/no-static-work>,
but that policy will now be more directly enforced.

Bug: 319647875
Test: atest SystemUiRoboTests
Test: atest SystemUiRavenTests
Change-Id: I5ef7a2ada3047d468ff335511467f9691752e9f1