Remove recoverBuilder use from test
We removed the usage of recoverBuilder in contentDescForNotification,
but left it in the test. This change removes it from the test as well,
and makes the notification passed to contentDescForNotification non-null
since there's no point in getting a description for a null notification.
I checked the callsites and this should be safe to do.

Bug: 281629927
Test: NotificationContentDescriptionTest
Flag: NONE
Change-Id: I1cea7d0163b5cd87ff3daf159cbec617b335d30d