Add -a to capture all the physical displays.
when -a is given, it will append a postfix to the FILENAME, such that
FILENAME_0.png, FILENAME_1.png. this won't break the existing behavior
as it doesn't mutate the name if it's taking a single display.
when both -a and -d are given, it will ignore -d and capture all the
displays.
also updated the usage doc to clarify it captures the default display
when the id is not given.

Test cases:
- no argument > prints out in the cmd window
- only -p > prints out in the cmd window as png format
- -h > correctly shows the usage
- -d with and without .png > saves as a file
- -a with and without .png > saves as files
- -d and -a > ignores -d and shows the same results as single -a

Bug: 321278149
Test: adb shell screencap with various flags
Flag: NA
Change-Id: Iecfeec1a1edbc95d7e8931ac3b22ac7f0706c3e7

Former-commit-id: e644f58c06768d19b1d6b2be4dd890c374bb3047