Fix NPE in LockscreenShadeTransitionController
When using scene container, QS is null. We add nullability to the
variable so we can handle that case. In most cases, the calls are not
necessary anymore because all the state animation is handled by STL, but
in the case we need a return value, we use `QSSceneAdapter` to retrieve
it.

Test: manual, no crash when going to lockscreen
Test: atest com.android.systemui.statusbar
Bug: 280887232
Flag: None
Change-Id: Iad0a49b71b2e3dad00193b47eed3985baa28f51f