Mark current tiles as auto-added
If a tile is currently in the list of tile but could be auto-added if
not tracked, mark it as auto-added.

Note that this only applies to tiles that can be auto-added only once
(AutoAddTracking.IfNotAdded), as a way to indicate that the user has
already seen the tile (and possibly may have removed it). This doesn't
apply to tiles that are always tracked (AutoAddTracking.Always), as
those can be auto-removed and auto-added again when the feature is
enabled again.

Test: atest AutoAddInteractorTest
Fixes: 324960944
Flag: ACONFIG com.android.systemui.qs_new_pipeline NEXTFOOD
Change-Id: I6ca4261ed98548549ad257d0ea9e9af6f3c1c8ba