fix(high contrast text): draw contrasting backgrounds when text changes color via Spans
We take into account the CharacterStyle spans that change the color of
the text to be different than the original global Paint color. This
matches the high contrast background to each span of text that is a
different color.

Also, fix the isHighContrastTextDark() function which was converting
LAB wrong because of incorrect documentation (see b/343778621)

Fix: 340552436
Flag: ACONFIG com.android.graphics.hwui.flags.high_contrast_text_small_text_rect TRUNKFOOD
Test: atest frameworks/base/core/tests/coretests/src/android/text/LayoutTest.java
 && atest frameworks/base/core/tests/coretests/src/android/text/SpanColorsTest.java

Change-Id: I8e9723409b8219ad9c869dc5a7c767d2ec669939

Former-commit-id: 48a980e116d4fe296d331c2226cd53c9f0fb5761