Pre-fetch fill request for credman
If by the time autofill manager client is notified of a credential view
with eligible credman request, then start pre-fetching fill response.
When the fill response is returned, the fill dialog will still be
suppressed during the display phase. If the view does not contain an
eligible credman request, then the pre-fetch won't start since it is
vulnerable to race condition.

Bug: 327032807
Test: locally, atest cts

Change-Id: I58c8fe6b71e3e882ed06ad6f5628f42a2a2cb291