[CarrierText] Update log messages for onSimStateChanged
Previously, CarrierTextManager was (incorrectly) claiming to be updating
carrier text when the onSimStateChangedCallback happened. In reality, we
only updated the text when the error state changed for that SIM.

This change logs a verbose line for every time the callback triggers,
and adds the slotId and simState to the line. As well as it updates the
REASON_ON_SIM_STATE_CHANGED to be REASON_SIM_ERROR_STATE_CHANGED to
better reflect intent.

Test: adb shell -n dumpsys activity service com.android.systemui/.SystemUIService CarrierTextManagerLog
Bug: 321918644
Flag: NONE
Change-Id: I068551bdfe97af556183c2041752b997df1714d4