Reland "Add support for generated previews to People Space widget"
This change updates PeopleSpaceWidgetManager to use
AppWidgetManager.setWidgetPreview to provide widget picker previews.
PeopleSpaceWidgetManager will set the previews when the user is first
unlocked.

This reland includes the fix from ag/26588265, which checks if the
provider is enabled before setting the preview. The provider may be
disabled on TvSystemUI

Bug: 308041327
Test: PeopleSpaceWidgetManagerTest#testUpdateGeneratedPreview_*
Flag: ACONFIG android.appwidget.generated_previews TRUNKFOOD
Change-Id: I475f35521764813348164847f1173bcbfc773de3