Camera2: Fix exception when cameraservice is disabled
When cameraservice is disabled on a device, Camera2 APIs should
treat it as if there are no camera devices enumerated and throw
the appropriate exceptions.

CameraManager's CameraDeviceSetup methods treated cameraservice
being disabled separately from no camera devices being enumerated.
This CL fixes the API behavior of #isCameraDeviceSetupSupported
and #getCameraDeviceSetup to throw an IllegalArgumentException
when cameraservice is disabled.

Bug: 325760092
Test: atest CtsCameraTestCases:CameraDeviceSetupTest passes on
      devices with cameraservice disabled.
Change-Id: I67ca0f0831846568d34a84df98b03b4f37d8b1ff