NSSL: Apply skinny notifs based on split shade state
Currently, we're relying on some complex config/resource logic to decide
whether to use the skinny notif layout, and it seems like it might be
flaky.

Therefore, check *directly* whether NSSL should use the split shade, and
use *that* to decide whether to use the skinny notif layout.

Bug: 328588062
Test: postsubmit
Flag: NA
Change-Id: Iaeb97db2e4808cbec131019a2f2dcc6a679fa642

Former-commit-id: 13203c2c38032acdc85f4927922179712ff60bc4