Camera: Skip extension processing if session is not valid
The preview request handler can run in a client supplied executor.
The corresponding thread can continue to be active and process
result callbacks after the camera device and session are no longer
valid which can have effect on the pending images.
To avoid accessing closed images, check whether the session is
still initialized.

Bug: 311634576
Test: atest -c -d cts/tests/camera/src/android/hardware/camera2/cts/CameraExtensionSessionTest.java
Change-Id: Ic6079f4eddc07a58b82039c59b46c1a1166a2c12