Perform bounds snap when not changing windowing mode.
ag/26334075 caused caption to not snap on screen when dragged into
status bar region but not in the TO_FULLSCREEN drag zone. A subsequent
corner resize would then show the task snapping back to its pre-drag
bounds. This would happen because no WCT would occur to finalize the
bounds as our assumption was a status bar drag wouldn't need one.

This CL fixes this by using snapTaskBoundsIfNecessary from the utility
class and setting bounds to the result in any drag that does not result
in a windowing mode change.

A handful of tests were removed from
Veiled/FluidResizeTaskPositionerTest as the functionality is no longer
present in those classes. DragPositioningCallbackUtilityTest covers
testing snap to valid drag area.

Bug: 328423886
Bug: 329013327
Test: Manual
Change-Id: I1cf7ca59ffb5f3600cd82213c62f02ae7f5b1878

Former-commit-id: 76e6f41a2a867267840d9d341507d3570da59ba5