Remove unnecessary query in IMMS#switchUserOnHandlerLocked()
This is a follow up CL to my previous CL [1], which started using

  InputMethodSettingsRepository

as a on-memory data repository of InputMethodSettings to avoid
unnecessary queries to PackageManagerService.

With this CL

  InputMethodManagerService#switchUserOnHandlerLocked()

also start using InputMethodSettingsRepository instead of querying
available IMEs for the new user dynamically.

This CL is supposed to improve the IME user switching latency.

 [1]: Ie2d3b7e9b75790f64789d0c2e96ae46b1dee3a1e
      62603d7e9979d0e6d510acfe10f0fc2aee461ef1 [formerly 03e0580e6f1ff8c4bb345747448b2569ebf7f1bf]

Bug: 309837937
Bug: 329163064
Test: presubmit
Change-Id: Icfdb247ee6bbcc43ce7b09954a85bfcff31c8441

Former-commit-id: 313a52758136c3661ec3ae2fe02abc50be8f3ffb