Change file access return types to ParcelFileDescriptor
FileInputStream works well with Java code, but for this usecase files may need to be passed onto JNI layer which requires access to fd-int and
FileInputStream or the underlying FileDescriptor type do not expose a
getter to the FD int which causes a lot of back and forth.

Changing to PFD and add checks alongside other validations in
system-server to ensure read only behaviour.

Test: cts in this topic added
Bug: 329656165
Change-Id: Ic5aa0d49e04afc74ff97833f3bbc5eab4790ef5b

Former-commit-id: 66abe4bcaf69d83f0ef64aff6d67b05715b48b3c