Skip sleep-token when switching display which will be on
To reduce unnecessary visibility or lifecycle changes.
That may improve the latency of display switch and avoid
showing intermediate content.

Note: mRequestedOrSleepingDefaultDisplay cannot be used for
this case because it is set to true by screenTurningOff but
it doesn't mean the device will sleep.

Maybe commit 2ca16286 and e6cf9f30 can be removed after
adoping this concept.

Bug: 303241079
Test: atest PhoneWindowManagerTests#testScreenTurnedOff
Test: Fold/unfold a foldable device.
      With Settings > Display > Continue using apps on fold
      - Keep screen on fold
        Home keeps resumed and won't receive invisible.
      - Turn off screen on fold
        Home can receive invisible and stopped.
Change-Id: If3848b4a6d67d71da7d735af56fcfd967152c1f6