Prevent pilfering of resize events.
Adds a check to see if drag resize handler should handle an event when
other caption input is being handled. This is used in to prevent
pilfering in view model when we want a corner drag.

This fixes an issue where drag events in the area where caption and
resize handles intersect would trigger both drag types but immediately
cancel the drag resize, resulting in a reposition that sets bounds to
fullscreen on ACTION_UP.

This also restores the functionality of resize handles to be slightly
inside caption bounds.

Bug: 328413639
Test: Manual
Change-Id: I56a5ab2c22c4d32f3eac3ec745820d18cfe83b73

Former-commit-id: 4f228c305d26ffba649565e780892e1963881039