Add UserHandle to MediaRouterService#showMediaOutputSwitcher
This change is preliminary work towards associating the output switcher
to a specific user id. Before this work, the OutputSwitcher is only
associated to a package name. With this change we are plumbing the user
handle (in addition to the existing package name) through the MediaRouter
service to SystemUI.

After this change, we need to plumb the user handle back from SystemUI
into MediaRouter2 (through SettingsLib), in parallel to what we already
do for the package name.

Bug: b/279555229
Test: Presubmit. No functional changes.
Flag: N/A as no logic added, just methods' signatures changes
Change-Id: I88e40b4f79807358c982d45e4930733d56d5f8a3

Former-commit-id: 08a45fcca24ce53e23f812e5899cf7f72354f5ef