When the DeviceEntryUdfpsRefactor is enabled, use an empty lock icon view controller
There's no need for the lock icon view controller's logic when the
device entry udfps refactor is enabled. When the flag is enabled,
SysUI relies on the DeviceEntryIconViewBinder + ViewModels for the
lock icon logic. This saves us from storing a the lock icon view
when we don't even use it.

Bug: 325766190
Test: android.platform.test.scenario.sysui.lockscreen.LockscreenAodTransitionMicrobenchmark#testLockscreenAodTransition
Flag: ACONFIG com.android.systemui.device_entry_udfps_refactor TEAMFOOD
Change-Id: I4edf4638b61135bd7764ae552b5b39576131ccb5

Former-commit-id: 5ceed31bba41f39859bda18318da39c730b218b5