Skip translating the x coordinate of the large clock
It turns out this wasn't a pivot problem like we thought. Instead the
large clock shouldn't be translated in the x direction. Comparing to the
version in KeyguardStatusViewController & KeyguardClockSwitchController
shows that only the small clock is translated in that direction.
Everything is still translated in the y direction, as that translation
was applied at the parent level directly to KeyguardStatusView.

Bug: 322199646
Test: Manually verified translation matched unflagged translation
Flag: ACONFIG com.android.systemui.migrate_clocks_to_blueprint STAGING
Change-Id: I27a6afb38538e128c600bcbd0c324fbad3b3e84c

Former-commit-id: 7db86120ffd32ad23644df2b5993873555efba12